requirements.txt
pdfplumber
camelot-py[cv]
tiktoken
faiss-cpu
google-generativeai
python-dotenv
streamlit
pandas
requests
